requirements.txt
pyautogui
opencv-python
pillow
numpy